cdk/lambda/spoken_llm/requirements.txt
python-socketio[asyncio_client]
pyaudio
aws-sdk-bedrock-runtime